[Cleanup] Add boundary checks for accessing UserVar
event->BaseVarIndex is set only at a few places in the code, so adding checks there is enough to make sure all uses of event->BaseVarIndex are safe.

Especially when computing a userVarIndex from event values (user input) the range should be checked before using it to access an element in the UserVar array.