[WiFi] Only allow automatic AP mode in first 5 minutes from boot
The timer to disable AP mode should be set as soon as the AP is started, to make sure it will be disabled.
It only will be increased when a client connects to the AP mode.

When disabling the AP mode, it needs a check to see if the AP mode was actually disabled before clearing the AP off timer.