Moved GPIO commands in the core
Hi all,
moved the GPIO commands for Internal pins, PCF and MCP in the core code from PLUGIN_WRITE.

Created several support functions to avoid repetition of the code.

The old code is still commented in order to simplify the debug and the comparison between the old code and the new changes.

Fixed some bugs along the way.

As the changes are really substantial, I really would like the community to test before releasing.