[Rules] Add check for correct received POST data when saving rules
I  added a very basic JavaScript line to the rules page, where it appends the length of the rules.
Only problem is that the length in JavaScript is with a newline counted as \n and it outputs to the node with a newline as \r\n, so there is an offset of 1 byte per line.
But still gives me a good indication whether data is lost, which happens quite a lot when you cross the 1600 bytes

Known issue:
If you try to save and it fails, then the stored rules will be shown (not the one you just tried to submit)