[SGP30] Quick update + Store baseline info for quick init
The sensor cannot output useful values after the IAQ init.
So we now keep track of the last known baseline values and restore them after the sensor is initialized.
For example after a reboot or a crash.

Also the values of TVOC = 0 and eCO2 = 400 will be ignored as new values right after sensor init.