[Cron] Fix handling bogus last/next times + restructure code
The plugin was assuming the values stored in UserVar were correct.
After a cold boot, or when the NTP time (or via GPS) was not yet set, these values for last time or next time could be just about any value.
This means it was possible the Cron task would never fire an event if the next time was computed far in the future.

Also restructured the code a bit, to make it more readable.