[Test] Add generated count output for sysinfo and CUL reader for debug
In order to test for missing messages when sending data to controllers.
The sysinfo and CUL reader plugin can generate a counter value which gets incremented on every call to PLUGIN_READ.
For the sysinfo plugin, this can also be used to output upto 4 'count' messages with the same value on each call to PLUGIN_READ.

The generated output of the CUL reader plugin will generate a string output starting with the counter, followed by N times the last decimal of the counter.
The total length of the generated string is as long as set in the task config.