[P036_FramedOLED] Reduce stack and heap memory usage
During load, the (large) task data struct was allocated on the heap, regardless whether it was already present in memory when the task was initialized.

During save, the (large) task data struct was allocated on the stack.

Both load and save now check whether there is already an instance present and use that instance.
If not, it creates a temporary instance for loading and saving of the settings.

Also the scrolling variable was not initialized, so it could happen the display would not start to show the pages as it was marked as "still being scrolling"