[TSL2591] Check sensor before reading + speedup
The old implementation did perform 3 readings, which depending on the set integration time could take a while.
But since all output values should be based on the same reading anyway, it is simply wrong to perform separate readings.

Also added a call to `begin()` of the library to make sure the gain and timing are correctly set.
This could cause incorrect readings if the sensor somehow got reset.

Fixes: #3347