[Dallas] Reschedule at normal interval to prevent drift caused by reading
Reading a sensor value from a Dallas sensor may take upto 800 msec, based on the set resolution.
This means the set interval for producing a sample will drift by roughly 800 - 900 msec per reading.
By recording the time when starting the reading and using that to reschedule the task timer this drift is minimized.