[GPIO] Only log errors when using positive GPIO
As discussed here: https://github.com/letscontrolit/ESPEasy/issues/3396#issuecomment-734434852

P029 Domoticz MQTT helper does allow to not use a GPIO pin.
The pin number is then set to -1.
Typical use case is to only act on the event generated by the helper plugin.
Therefore it is a valid situation and should not pollute the logs.