[Dallas] Fix Dallas sensor not being read if optional pin unset
A recent feature addition allowed for an optional separation of RX and TX pin for using Dallas sensors on a Shelly with an isolator board.
This still works with using only 1 pin, but not if you set the optional pin to "None".

This fixes this inconvenience by just copying the optional pin nr from the standard pin when initializing the sensor. (it is not duplicating it upon saving as that may be confusing)