[ESPEasy-NOW] Use hidden ESPEasy-now AP for stability
For stability of ESPEasy-NOW protocol it is better to connect to a virtual AP hosted by some other ESPEasy-NOW node.
We don't send data over it, but it prevents some very nasty (undocumented) power management in the closed source wifi libraries of Espressif to kick in and mess up the ESPEasy-NOW layer.