[ESPEasy-NOW] Let traceroutes expire & only route on traceroutes
Routes advertised in "discovery" packets are of no use for routing.
But they are useful for showing the nodes state on the root page of the web UI.
So the routes received via traceroute are now used for routing and those received via discovery packets are used for viewing/debugging purposes only.

A traceroute will be valued on the quality of the hops and the number of hops.