[ESPEasy-NOW] Add peer manager to help broadcast packets succeed
ESP-now does send broadcast packets to known peers (and those that still may receive it).
This makes broadcast packets much more reliable when the intended recipient is a known peer.
However we only have a limited number of peer slots (depending on whether or not an encryption key is used), so we must keep track of 'popular' nodes to keep those active as peer.