[WiFi] Fix not able to reconnect if unit immediately disconnects
Sometimes the ESP may get in some strange loop where it immediately gets disconnected.
Often with a disconnect reason like "4 way handshake timeout".
This may be impossible for the ESP to get out of, sometimes even a warm reboot (crash) may not help.
The only way to get out of it is a full WiFi disconnect.