[Provisioning] Allow to fetch config via commands from HTTP server
To ease provisioning a number of nodes, an administrator should be able to configure some host + credentials to fetch files from a server and allow to update a node via rules.

The commands do not allow to fetch files from just anywhere, since the user must explicitly enable and configure it.
Therefore it is controlled by the administrator whether a file may be fetched and what file to fetch.

It is also possible to include some system variable in the URL, like `%mac%` which will be URL encoded when fetching the file.