[GPIO Selection] Filter possible GPIO pins based on needed properties
When selecting a GPIO pin from the web UI, one should only be allowed to select pins that can be used.
A pin will be disabled for selection if it is already used on I2C, serial, SPI or ethernet.
It is also displayed next to the GPIO description.

Pins that ar not usable at all, for example if you need an output pin, it makes no sense to show input-only pins on an ESP32.
Generating the pin selection combobox now also does not need to be allocated in memory anymore as it is streamed directly to the client.