[ESP32] Disable DHCP server on fake AP to suppress logs to Serial0
See: [invalid-event-when-client-connect-ap-debug-output](https://community.platformio.org/t/invalid-event-when-client-connect-ap-debug-output/11539)
Apparently the ESP32 IDF code does not have a wrapper around this log entry so it is logging it always regardless of the desired settings.

What happens is that we're having a hidden AP active and if it is scanned, it might output this log.