[Cache Controller] Update dump.htm to not check for valid timestamp
When used on a node that cannot have its time set to a proper time, the `dump5.htm` would not export any data in the export CSV file.
This because of a very simple check for a valid time stamp where the Unix Time had to be > 1.5 billion.
However on a node which doesn't have any time set, you can't check for this.