[ESPEasySerial] Fix issues on ESP32 with serial (restart serial)
ESP32 does have an issue with restarting serial communication.
Calling `end()` followed by `begin()` on the serial port, may render the port unusable.

This work-around is to keep track of the last used pins and do not call `end()` and `begin()` if the pins have not changed.

It is a really dirty hack, but it does work.

The serial port may be stopped/started when updating a setting on a task, so this was for sure a big issue on ESP32.