[P016] Accept decode_type_t::UNKNOWN as valid IR code
In order to use cheap Chinese IR remote controls accept IR code that can not be decoded by the IR library funtions as valid IR code for P016 commands.
An additional checkbox is added to the web interface of the [P016] device.
If this checkbox is checked the decode_type is changed to RAW before calling P016_data->AddCode() or  P016_data->ExecuteCode().