[ETH] Turn off Ethernet power on reboot + add delays for init LAN chip
Having a power pin for the Ethernet chip appears to be a bit timing critical on some boards.
This mainly has to do with tolerances in capacitors.
We have to make sure the Ethernet chip has enough time before calling the ETH.begin() function to get ready or else initialization may fail.

Also on boards which do have an external crystal, we must make sure the device is turned off when rebooting or else this clock signal on GPIO-0 may put the node into flash mode at boot. (this is theoretical as a warm boot does not appear to be checking the state of GPIO-0)