[Fonts] Only include font data once in built bin file.
With globally defined char arrays in .h files, those large arrays will be linked multiple times if that .h file is included more than once.

Now the arrays are in a .cpp file with a simple get function to access them. This prevents the array to appear multiple times, as did happen at least on ESP32 MAX builds.