[Framed OLED] Reduce memory usage, especially during load/save
No need to keep the raw data structure in memory, as this structure reserves the maximum allowed string length, where often not all strings are used and for sure not all at their max. length.

When loading/saving, only a small part of this raw structure needs to be active in memory.