[Dallas] Add retry to init read or collect data from temp sensor
It appears most of the read errors occur on the first sensor of a task.
So perhaps there is some kind of disturbance on the line or maybe there is some kind of race condition when switching tasks.
This means it can be useful to retry to  start a temperature reading on a sensor or retry to collect data from a sensor as it doesn't happen often (or at least should not happen often)