[HLW8012] Improve stability and resolution Voltage/Current measurement
Every now and then there would be some fluke measurements, due to some late interrupt handling.
Now the pulses on CF1 are counted as well as measured.
For low pulse counts, only the duration of the last pulse is taken as it may take some time to get stable pulses. Thus then it is better to use only the last one of a period.
For high pulse counts, it is better to take the duration over a longer time and then average over the nr of pulses.

Before this, the "resolution" of the mains voltage measurements were roughly 8V and thus completely unusable.
With these changes, the resolution of the mains voltage is closely following the measured voltage of the CSE7766, used in the Sonoff POW2. (less than 0.1 V off in the same wall socket)