[Setup] Fix CSS on setup page + WiFi/Eth event handling
On ESP32, the event handling for WiFi (and Ethernet) needs to be re-registered when interfaces are turned on/off.
Otherwise events like connect to ESP AP will not be registered and thus the captive login page will not be served.

Also Favicon and CSS were not properly shown on the setup page on ESP32/ESP32-S2 due to a mismatch in file name comparison.