[Cleanup] Do not call own class functions from constructor/destructor
A constructor or destructor should not call functions of its own class.
Especially not when these classes use inheritance.

Problem is that inside the constructor a class is not yet completely initialized.
Accessing non virtual methods can be done, but you better should not do it as it may lead to subtle bugs when such functions are later changed.