[ESP32] Fix ECO mode on ESP32
The option to force WiFi no sleep was unavailable on ESP32.
This effectively made the ECO mode useless on ESP32.
Now this is available again, the current consumption will reduce when enabling ECO mode.
However the current consumption can be lowered even more by switching the ESP32 to 80 MHz.
This will only be set at boot, as dynamic frequency switching is a bit funky.
However Espressif's dynamic frequency scaling does seem to work well, so this is also enabled at boot when ECO mode is set.

The WiFi TX power settings, which are present on ESP8266, are still disabled on ESP32 as it does result in less reliable WiFi connection and higher current consumption compared to the dynamic wifi light sleep mode.