[ESP32 ETH] Fix setting DNS on latest ESP32 SDK
When switching off WiFi, the DNS server records are cleared.
This is checked for and a cached version is restored.

However, on the latest ESP32 SDK, IPv6 has been enabled which changed checking for an IPAddress to be set.
Thus restoring a cached version could simply reset any existing valid entry.