[Serial Helper] Disable updating GPIOs on switch serial port on webpage
Toggling the serial port was updating the selected RX and TX.
However this was also done on load of the page, so any selected pins would then be reset to the defaults for that port.
Thus when not changing something and pressing "Submit" would change the set GPIO pins for the serial port.