[GPS] Improve GPS resolution
GPS is one of the very few use cases, where the resolution of a float is not enough.
A float only has a resolution of 6 decimals.
However, when you are located at a location where the latitude or longitude requires 2 or 3 decimal places to represent the degrees, not a lot of resolution is left to pinpoint the actual position.

Before this fix, the standard deviation of my position (53 degree North, 6 East) was about 15 - 20 meters on a stationary GPS module.
With this fix, the standard deviation is about 5 meter for the same stationary GPS modules.